+++ homepage/tests/test_page_index.py
@@ -1,9 +1,6 @@
 """
-Includes all unit and integration tests for the homepage app.
+Includes tests to test the index page of the homepage app.
 """
-from django.contrib.auth.models import User
-
-# pylint: disable=unused-import
 from django.test import TestCase, tag
 from django.urls import reverse
 from django.utils.translation import gettext_lazy
@@ -13,7 +10,7 @@ from django.utils.translation import gettext_lazy
 @tag("page-index")
 class TestIndexPage(TestCase):
     """
-    Including all tests for the index page.
+    Includes tests to test the index page of the homepage app.
     """
 
     @tag("template")
@@ -61,51 +58,3 @@ class TestIndexPage(TestCase):
             f'{reverse("homepage:index", kwargs={"active_category": "physics"})}">Physics</a>'
         )
         self.assertContains(response, find_physics, html=True)
-
-    @tag("login")
-    def test_login_ability(self) -> None:
-        """
-        Test if you can log into in a general manner. It does not test the login page.
-        """
-        user = User.objects.create(username="testuser")
-        user.set_password("12345")
-        user.save()
-
-        self.assertTrue(self.client.login(username="testuser", password="12345"))
-
-    @tag("login", "routing")
-    def test_login_page_200(self) -> None:
-        """
-        Test if the login page is available
-        """
-        response = self.client.get(reverse("auth:login"))
-        self.assertEqual(response.status_code, 200)
-
-    @tag("login", "routing")
-    def test_login_page_redirect(self) -> None:
-        """
-        Test if the login page is available and redirects successfully
-        """
-        user = User.objects.create(username="testuser")
-        user.set_password("12345")
-        user.save()
-
-        response = self.client.post(
-            reverse("auth:login") + "?next=/", {"username": "testuser", "password": "12345"}, follow=True
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, "homepage/index.html")
-
-    @tag("logout", "routing")
-    def test_logout_page_redirect(self) -> None:
-        """
-        Test if the logout page is available and redirects successfully
-        """
-        user = User.objects.create(username="testuser")
-        user.set_password("12345")
-        user.save()
-
-        self.assertTrue(self.client.login(username="testuser", password="12345"))
-        response = self.client.get(reverse("auth:logout") + "?next=/", follow=True)
-        self.assertEqual(response.status_code, 200)
-        self.assertTemplateUsed(response, "homepage/index.html")
